+++ src/core.js
@@ -15,7 +15,7 @@
 var Plotly = require('./plotly');
 
 // package version injected by `npm run preprocess`
-exports.version = '1.18.1';
+exports.version = '1.19.1';
 
 // inject promise polyfill
 require('es6-promise').polyfill();