+++ src/traces/contour/make_color_map.js
@@ -10,9 +10,7 @@
 'use strict';
 
 var d3 = require('d3');
-
-var getColorscale = require('../../components/colorscale/get_scale');
-
+var Colorscale = require('../../components/colorscale');
 
 module.exports = function makeColorMap(trace) {
     var contours = trace.contours,
@@ -22,7 +20,7 @@ module.exports = function makeColorMap(trace) {
         nc = Math.floor((end + cs / 10 - start) / cs) + 1,
         extra = contours.coloring === 'lines' ? 0 : 1;
 
-    var scl = getColorscale(trace.colorscale),
+    var scl = trace.colorscale,
         len = scl.length;
 
     var domain = new Array(len),
@@ -69,10 +67,10 @@ module.exports = function makeColorMap(trace) {
         }
     }
 
-    var colorMap = d3.scale.linear()
-        .interpolate(d3.interpolateRgb)
-        .domain(domain)
-        .range(range);
-
-    return colorMap;
+    return Colorscale.makeColorScaleFunc({
+        domain: domain,
+        range: range,
+    }, {
+        noNumericCheck: true
+    });
 };